firebase/functions/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
blinker==1.9.0
cachecontrol==0.14.3
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cloudevents==1.11.0
cryptography==46.0.1
deprecation==2.1.0
distro==1.9.0
firebase-admin==7.1.0
firebase-functions==0.4.3
flask==3.1.2
flask-cors==6.0.1
frozenlist==1.7.0
functions-framework==3.9.2
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.0
google-crc32c==1.7.1
google-events==0.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.75.1
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
jiter==0.11.0
markupsafe==3.0.3
msgpack==1.1.1
multidict==6.6.4
openai==1.109.1
packaging==25.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic-core==2.33.2
pyjwt==2.10.1
pyyaml==6.0.3
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
starlette==0.48.0
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.37.0
uvicorn-worker==0.4.0
watchdog==6.0.0
werkzeug==3.1.3
yarl==1.20.1
